requirements.txt
streamlit
pandas
gspread
python-dotenv
google-auth
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-groq
sentence-transformers
faiss-cpu